Persona: mergeStyles part 3, split sub-components (#4496)
* Split PersonaCoin

* Split PersonaPresence

* Move to sub-folders
