+++ common/changes/office-ui-fabric-react/mergeStyles-persona-part3_2018-04-09-17-43.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "office-ui-fabric-react",
+      "comment": "Split PersonaCoin and PersonaPresence files for more consistency.",
+      "type": "patch"
+    }
+  ],
+  "packageName": "office-ui-fabric-react",
+  "email": "jordancjanzen@gmail.com"
+}
