+++ common/changes/office-ui-fabric-react/mergeStyles-persona-part3_2018-04-09-20-37.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "office-ui-fabric-react",
+      "comment": "Move sub-components to sub-folders",
+      "type": "patch"
+    }
+  ],
+  "packageName": "office-ui-fabric-react",
+  "email": "jordanjanzen@jordanjanzen.com"
+}
